divide tests directory into two directories un-peg versions in requirements_dev.txt
