+++ requirements_dev.txt
@@ -1,8 +1,6 @@
-pip==9.0.1
-wheel==0.30.0
-flake8==3.5.0
-tox==2.9.1
-coverage==4.5.1
-mkdocs==0.17.2
-pytest==3.4.2
-pytest-runner==4.0.0
+wheel
+flake8
+tox
+coverage
+mkdocs
+pytest
