+++ setup.py
@@ -1,9 +1,6 @@
 # !/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-# Note: To use the 'upload' functionality of this file, you must:
-#   $ pip install twine
-
 import io
 import os
 import re
@@ -19,10 +16,7 @@ AUTHOR = 'Brian Okken'
 REQUIRES_PYTHON = '>=3.6.0'
 VERSION = '0.1.4'
 
-# What packages are required for this module to be executed?
-REQUIRED = [
-    'Click==6.7', 'tinydb==3.8.0', 'attrs==17.4.0'
-]
+REQUIRED = [ 'Click==6.7', 'tinydb==3.8.0', 'attrs==17.4.0' ]
 
 # The rest you shouldn't have to touch too much :)
 # ------------------------------------------------
@@ -32,7 +26,6 @@ REQUIRED = [
 here = os.path.abspath(os.path.dirname(__file__))
 
 # Import the README and use it as the long-description.
-# Note: this will only work if 'README.rst' is present in your MANIFEST.in file!
 with io.open(os.path.join(here, 'README.rst'), encoding='utf-8') as f:
     long_description = '\n' + f.read()
